Second Revision

+++ main.py
@@ -1,9 +1,10 @@
 import os
 import hmac
 import re
-import random
-from string import letters
-import hashlib
+
+from post import Post
+from comment import Comment
+from user import User
 
 import jinja2
 import webapp2
@@ -78,78 +79,22 @@ def render_post(response, post):
     response.out.write(post.content)
 
 
-# user info
-def make_salt(length=5):
-    return ''.join(random.choice(letters) for x in xrange(length))
-
-
-# makes a the pw hashed and h is stored to db
-def make_pw_hash(name, pw, salt=None):
-    if not salt:
-        salt = make_salt()
-    h = hashlib.sha256(name + pw + salt).hexdigest()
-    return '%s,%s' % (salt, h)
-
-
-# takes name and pw and checks values in db
-def valid_pw(name, password, h):
-    salt = h.split(',')[0]
-    return h == make_pw_hash(name, password, salt)
-
-
-# creates ancestor element to store all users in db
-def users_key(group='default'):
-    return db.Key.from_path('users', group)
-
-
-# user object stored in db
-class User(db.Model):
-    name = db.StringProperty(required=True)
-    pw_hash = db.StringProperty(required=True)
-    email = db.StringProperty()
-
-    # convenience function for looking up user by id
-    @classmethod
-    def by_id(cls, uid):
-        return cls.get_by_id(uid, parent=users_key())
-
-    # convenience function for looking up user by name
-    @classmethod
-    def by_name(cls, name):
-        u = cls.all().filter('name =', name).get()
-        return u
-
-    # convenience function for creating new user
-    @classmethod
-    def register(cls, name, pw, email=None):
-        pw_hash = make_pw_hash(name, pw)
-        return cls(parent=users_key(),
-                   name=name,
-                   pw_hash=pw_hash,
-                   email=email)
-
-    # convenience function for making sure
-    # user exists and pw is valid
-    @classmethod
-    def login(cls, name, pw):
-        u = cls.by_name(name)
-        if u and valid_pw(name, pw, u.pw_hash):
-            return u
-
-
 # Registration
-
 USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
-PASSWORD_RE = re.compile(r"^.{3,20}$")
-EMAIL_RE = re.compile(r"^[\S]+@[\S]+.[\S]+$")
 
 
 def val_username(username):
     return username and USER_RE.match(username)
 
 
+PASSWORD_RE = re.compile(r"^.{3,20}$")
+
+
 def val_password(password):
     return password and PASSWORD_RE.match(password)
+
+
+EMAIL_RE = re.compile(r"^[\S]+@[\S]+.[\S]+$")
 
 
 def val_email(email):
@@ -209,7 +154,7 @@ class Register(Signup):
             # if user is valid then user info is
             # stored to db and user redirected to main pg
             self.login(u)
-            self.redirect('/')
+            return self.redirect('/')
 
 
 class Login(Handler):
@@ -225,7 +170,7 @@ class Login(Handler):
         u = User.login(username, password)
         if u:
             self.login(u)
-            self.redirect('/')
+            return self.redirect('/')
         else:
             msg = 'Invalid login'
             self.render('login.html', error=msg)
@@ -242,39 +187,6 @@ def blog_key(name='default'):
     return db.Key.from_path('blogs', name)
 
 
-class Post(db.Model):
-    subject = db.StringProperty(required=True)
-    content = db.TextProperty(required=True)
-    created = db.DateTimeProperty(auto_now_add=True)
-    last_modified = db.DateTimeProperty(auto_now=True)
-    created_by = db.TextProperty()
-    likes = db.IntegerProperty(required=True)
-    liked_by = db.ListProperty(str)
-
-    @classmethod
-    def by_post_name(cls, name):
-        # find post by name as name
-        u = cls.all().filter('name =', name).get()
-        return u
-
-    def render(self):
-        self._render_text = self.content.replace('\n', '<br>')
-        return render_str("post.html", p=self)
-
-    @property
-    def comments(self):
-        return Comment.all().filter("post =", str(self.key().id()))
-
-
-class Comment(db.Model):
-    comment = db.StringProperty(required=True)
-    post = db.StringProperty(required=True)
-
-    @classmethod
-    def render(self):
-        self.render("newcomment.html")
-
-
 class MainPage(Handler):
     def get(self):
         posts = Post.all().order('-created')
@@ -288,8 +200,7 @@ class PostHandler(Handler):
         print post
 
         if not post:
-            self.error(404)
-            return
+            return self.error(404)
 
         self.render("permalink.html", post=post)
 
@@ -309,11 +220,11 @@ class NewPost(Handler):
         if self.user:
             self.render("newpost.html")
         else:
-            self.redirect("/login")
+            return self.redirect("/login")
 
     def post(self):
         if not self.user:
-            self.redirect('/')
+            return self.redirect('/')
 
         subject = self.request.get('subject')
         content = self.request.get('content')
@@ -330,8 +241,8 @@ class NewPost(Handler):
             print "post created by", n1
         else:
             error = "oops! you need a subject and content!"
-            self.render("newpost.html", subject=subject, content=content,
-                        error=error)
+            return self.render("newpost.html", subject=subject,
+                               content=content, error=error)
 
 
 class UpdatePost(Handler):
@@ -343,32 +254,32 @@ class UpdatePost(Handler):
             post = db.get(key)
             n1 = post.created_by
             n2 = self.user.name
-            print "n1 = ", n1
-            print "n2 = ", n2
             if n1 == n2:
                 key = db.Key.from_path('Post', int(post_id), parent=blog_key())
                 post = db.get(key)
                 print "post = ", post
                 error = ""
-                self.render("updatepost.html", subject=post.subject,
-                            content=post.content, error=error)
+                return self.render("updatepost.html", subject=post.subject,
+                                   content=post.content, error=error)
             else:
-                self.redirect("/editDeleteError")
+                return self.redirect("/editDeleteError")
 
     def post(self, post_id):
         if not self.user:
             self.redirect("/login")
-        else:
-            subject = self.request.get('subject')
-            content = self.request.get('content')
+
+        if post_id:
             key = db.Key.from_path('Post', int(post_id), parent=blog_key())
             p = db.get(key)
-            p.subject = self.request.get('subject')
-            p.content = self.request.get('content')
-            p.put()
-            self.redirect('/%s' % str(p.key().id()))
-            pid = p.key().id()
-            print "pid = ", str(pid)
+            n1 = p.created_by
+            n2 = self.user.name
+            if n1 == n2:
+                p.subject = self.request.get('subject')
+                p.content = self.request.get('content')
+                p.put()
+                self.redirect('/%s' % str(p.key().id()))
+                pid = p.key().id()
+                print "pid = ", str(pid)
 
 
 class LikePost(Handler):
@@ -382,12 +293,12 @@ class LikePost(Handler):
             current_user = self.user.name
 
             if author == current_user or current_user in post.liked_by:
-                self.redirect('/likeError')
+                return self.redirect('/likeError')
             else:
                 post.likes = post.likes + 1
                 post.liked_by.append(current_user)
                 post.put()
-                self.redirect('/')
+                return self.redirect('/')
 
 
 class DeletePost(Handler):
@@ -404,9 +315,9 @@ class DeletePost(Handler):
                 key = db.Key.from_path('Post', int(post_id), parent=blog_key())
                 post = db.get(key)
                 post.delete()
-                self.render("deletepost.html")
+                return self.render("deletepost.html")
             else:
-                self.redirect("/editDeleteError")
+                return self.redirect("/editDeleteError")
 
 
 class NewComment(Handler):
@@ -414,13 +325,12 @@ class NewComment(Handler):
         # Shows the new comment page
         if not self.user:
             error = "You must be logged in to comment"
-            self.redirect("/login")
-            return
+            return self.redirect("/login", error=error)
         post = Post.get_by_id(int(post_id), parent=blog_key())
         subject = post.subject
         content = post.content
-        self.render("newcomment.html", subject=subject, content=content,
-                    pkey=post.key())
+        return self.render("newcomment.html", subject=subject, content=content,
+                           pkey=post.key())
 
     def post(self, post_id):
         # If a ew comment was made,
@@ -428,21 +338,20 @@ class NewComment(Handler):
         key = db.Key.from_path('Post', int(post_id), parent=blog_key())
         post = db.get(key)
         if not post:
-            self.error(404)
-            return
+            return self.error(404)
         # make sure user is signed in
         if not self.user:
-            self.redirect('login')
+            return self.redirect('login')
         # creates the comment
         comment = self.request.get('comment')
         if comment:
             c = Comment(comment=comment, post=post_id, parent=self.user.key())
             c.put()
-            self.redirect('/%s' % str(post_id))
+            return self.redirect('/%s' % str(post_id))
         else:
             error = "please provide a comment!"
-            self.render("permalink.html", post=post,
-                        content=content, error=error)
+            return self.render("permalink.html", post=post,
+                               error=error)
 
 
 class UpdateComment(Handler):
@@ -450,29 +359,32 @@ class UpdateComment(Handler):
         post = Post.get_by_id(int(post_id), parent=blog_key())
         comment = Comment.get_by_id(int(comment_id), parent=self.user.key())
         if comment:
-            self.render("updatecomment.html", subject=post.subject,
-                        content=post.content, comment=comment.comment)
+            return self.render("updatecomment.html", subject=post.subject,
+                               content=post.content, comment=comment.comment)
         else:
-            self.redirect('/commenterror')
+            return self.redirect('/commenterror')
 
     def post(self, post_id, comment_id):
         comment = Comment.get_by_id(int(comment_id), parent=self.user.key())
         if comment.parent().key().id() == self.user.key().id():
             comment.comment = self.request.get('comment')
             comment.put()
-        self.redirect('/%s' % str(post_id))
+        return self.redirect('/%s' % str(post_id))
 
 
 class DeleteComment(Handler):
     def get(self, post_id, comment_id):
         post = Post.get_by_id(int(post_id), parent=blog_key())
-        # this ensures the user created the comment
+        # This ensures the user created the comment.
         comment = Comment.get_by_id(int(comment_id), parent=self.user.key())
         if comment:
-            comment.delete()
-            self.redirect('/%s' % str(post_id))
+            n1 = post.created_by
+            n2 = self.user.name
+            if n1 == n2:
+                comment.delete()
+                return self.redirect('/%s' % str(post_id))
         else:
-            self.redirect('/commenterror')
+            return self.redirect('/commenterror')
 
 
 class CommentError(Handler):
